requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
alembic
python-multipart
